backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
ffmpeg-python
moviepy
pillow
gTTS
python-dotenv